v. 19 adjust food direction priority calculation; switch to Double weights

+++ src/main/java/org/example/competition22/logic/BodyChecker.java
@@ -7,13 +7,13 @@ import org.example.competition22.data.MoveRequest;
 import java.util.Map;
 
 public class BodyChecker {
-    public static void check(MoveRequest request, Map<Direction, Integer> directions) {
+    public static void check(MoveRequest request, Map<Direction, Double> directions) {
         if (request.you.body.size() == 1) return;
         for (Direction direction : Direction.values()) {
             var possibleMove = Coordinate.getNextCoordinate(request.you.head, direction);
             for (Coordinate coordinate : request.you.body.subList(0, request.you.body.size()-1)) {
                 if (coordinate.equals(possibleMove)) {
-                    directions.put(direction, -1);
+                    directions.put(direction, -1.0);
                 }
             }
         }
